+++ src/Pipe.java
@@ -12,7 +12,7 @@ public class Pipe implements Collidable{
     }
 
     @Override
-    public double getX(){
+    public int getX(){
         return x;
     }
 
@@ -21,17 +21,17 @@ public class Pipe implements Collidable{
     }
 
     @Override
-    public double getY(){
+    public int getY(){
         return y;
     }
 
     @Override
-    public double getWidth(){
-        return image.getWidth();
+    public int getWidth(){
+        return (int)image.getWidth();
     }
 
     @Override
-    public double getHeight(){
-        return image.getHeight();
+    public int getHeight(){
+        return (int)image.getHeight();
     }
 }
